Move types and initcode functions into more specific directories
